+++ ex00/BitcoinExchange.hpp
@@ -6,14 +6,14 @@
 /*   By: tobourge <tobourge@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/28 13:16:01 by tobourge          #+#    #+#             */
-/*   Updated: 2025/10/30 11:19:25 by tobourge         ###   ########.fr       */
+/*   Updated: 2025/10/30 11:23:04 by tobourge         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #pragma once
 
 #include <string>
-#include <cstring>
+#include <string>
 #include <ctype.h>
 #include <iostream>
 #include <fstream>
@@ -24,7 +24,7 @@
 #include <algorithm>
 #include <ctime>
 
-typedef std::map<std::string, int> map_t;
+typedef std::map<std::string, float> map_t;
 
 
 bool        isFile(std::string filename);
